Only purge if the map isn't empty

+++ paper/src/main/java/com/badbones69/crazycrates/paper/support/holograms/types/DecentHologramSupport.java
@@ -37,6 +37,8 @@ public class DecentHologramSupport implements HologramManager {
 
     @Override
     public void purge() {
+        if (holograms.isEmpty()) return;
+
         holograms.forEach((key, value) -> value.delete());
         holograms.clear();
     }
